Moved remaining scattered audio classes to Audio folder.
